requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
alembic==1.13.3
httpx==0.27.2
apscheduler==3.10.4
opencv-python==4.10.0.84
numpy==1.26.4
pyyaml==6.0.2

Jinja2==3.1.4
python-multipart==0.0.9

icmplib==3.0

XlsxWriter==3.2.0
pytz==2024.1
